+++ src/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- Added compress() that accepts a float4 and uint4.
 - Added math.project() and math.projectsafe() for vector projection.
 - Added math.EPSILON, math.INFINITY, math.NAN and their double counterparts.
 - Added [Serializable] to RigidTransform.